Fixed FTUserCache issue, when data not being saved properly. Updated FTBaseView & FTBaseViewController: to support Interface builder allocation.

+++ FTMobileCoreUI/FTMobileCoreUI/BaseComponents/CoreComponents/FTWebView.swift
@@ -56,7 +56,7 @@ extension FTWebView {
         return "document.getElementsByTagName('body')[0]"
     }
     
-    public func setContentFontSize(_ size: Int) {
+    public func setContentFontSize(_ size: Float) {
         
         if size >= 10 {
             let js = self.getHTMLBodyText() + ".style.webkitTextSizeAdjust= '\(size)%'"
